Extract window stuff
