+++ src/bootstrap/customComponents.js
@@ -5,8 +5,8 @@ const Workspace = require('ui/workspace')
 const { WindowManager } = require('ui/windowManager')
 const Main = require('ui/main')
 const Utils = require('../utils')
-const { CustomWindowManager } = require('../customWindowManager')
-const { CustomWindowOverlay } = require('../customWindowOverlay')
+const { CustomWindowManager } = require('../window/customWindowManager')
+const { CustomWindowOverlay } = require('../window/customWindowOverlay')
 const { CustomWorkspaceView } = require('../customWorkspaceView')
 const { Search } = require('../search')
