+++ src/window/customWindowOverlay.js
