+++ src/window/selectedWindow.js
